All the methods are in the main. I just used one class to do everything, just for practicing.
Static:The static member can be accessed before the class object is
created.

A static variable is common to all the instances (or objects) of the
class because it is a class level variable.

Static methods in java belong to the class (not an instance of it). They
use no instance variables and will usually take input from the
parameters, perform actions on it, then return some result.

This project used TDD with Junit test